+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Added
+
+- File transfer strategy for Azure blob storage
+
 ## [0.224.0-rc.0] - 2023-06-04
 
 ### Authors